requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.0
openpyxl==3.1.2
xlrd==2.0.1
sqlalchemy==2.0.20
python-multipart==0.0.6
google-generativeai==0.3.1
pydantic==2.3.0
sqlite3==2.6.0 